chore: merge rough edges
